fix typo rename folder
