+++ services/web/shared/components/index.ts
@@ -2,4 +2,4 @@ export * from './UnderConstruction';
 export * from './basic';
 export * from './LoginPage';
 export * from './Header';
-export * from './AppWrapper';
+export * from './AppContent';
